requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
numpy
sqlalchemy
psycopg2-binary
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib
bcrypt==3.2.2
tensorflow-cpu==2.15.0
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai